Rename elements-metadata.ts to elements-format.ts
