+++ gulpfile.js
@@ -99,7 +99,7 @@ task('test', ['build'], () => {
 });
 
 task('json-schema', function() {
-  const inPath = 'src/elements-metadata.ts';
+  const inPath = 'src/elements-format.ts';
   const outPath = 'lib/analysis.schema.json';
   return gulp.src(inPath).pipe(newer(outPath)).pipe(shell([
     `./node_modules/.bin/typescript-json-schema --required ${inPath} Elements > ${outPath}.temp`,
