+++ src/test/analysis_test.ts
@@ -19,7 +19,7 @@ import * as path from 'path';
 import {Analysis, ValidationError} from '../analysis';
 import {Analyzer} from '../analyzer';
 import {ElementDescriptor} from '../ast/ast';
-import {Elements} from '../elements-metadata';
+import {Elements} from '../elements-format';
 import {generateElementMetadata} from '../generate-elements';
 import {FSUrlLoader} from '../url-loader/fs-url-loader';
