+++ src/test/elements-metadata_tests.ts
@@ -1,6 +1,18 @@
-import {Elements, Element} from '../elements-metadata';
+/**
+ * @license
+ * Copyright (c) 2015 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
 
-
+import {Element, Elements} from '../elements-format';
 
 // An example, the correct compilation thereof acts as a test.
 const paperButtonElement: Element = {
